Tests: Unify the test catalog setting
